remove unused db primary/replica/coordinator from bsky dataplane
