+++ packages/bsky/src/data-plane/server/background.ts
@@ -1,5 +1,5 @@
 import PQueue from 'p-queue'
-import { PrimaryDatabase } from './db'
+import { Database } from './db'
 import { dbLogger } from '../../logger'
 
 // A simple queue for in-process, out-of-band/backgrounded work
@@ -7,7 +7,7 @@ import { dbLogger } from '../../logger'
 export class BackgroundQueue {
   queue = new PQueue()
   destroyed = false
-  constructor(public db: PrimaryDatabase) {}
+  constructor(public db: Database) {}
 
   add(task: Task) {
     if (this.destroyed) {
@@ -32,4 +32,4 @@ export class BackgroundQueue {
   }
 }
 
-type Task = (db: PrimaryDatabase) => Promise<void>
+type Task = (db: Database) => Promise<void>
