+++ packages/bsky/src/data-plane/server/db/index.ts
@@ -1,3 +1,1 @@
-export * from './primary'
 export * from './db'
-export * from './coordinator'
