+++ packages/bsky/src/data-plane/server/indexing/plugins/list-block.ts
@@ -3,7 +3,7 @@ import { AtUri, normalizeDatetimeAlways } from '@atproto/syntax'
 import { CID } from 'multiformats/cid'
 import * as ListBlock from '../../../../lexicon/types/app/bsky/graph/listblock'
 import * as lex from '../../../../lexicon/lexicons'
-import { PrimaryDatabase } from '../../db'
+import { Database } from '../../db'
 import { DatabaseSchema, DatabaseSchemaType } from '../../db/database-schema'
 import RecordProcessor from '../processor'
 import { BackgroundQueue } from '../../background'
@@ -71,7 +71,7 @@ const notifsForDelete = () => {
 export type PluginType = RecordProcessor<ListBlock.Record, IndexedListBlock>
 
 export const makePlugin = (
-  db: PrimaryDatabase,
+  db: Database,
   background: BackgroundQueue,
 ): PluginType => {
   return new RecordProcessor(db, background, {
