+++ packages/bsky/src/data-plane/server/indexing/plugins/profile.ts
@@ -4,7 +4,7 @@ import * as Profile from '../../../../lexicon/types/app/bsky/actor/profile'
 import * as lex from '../../../../lexicon/lexicons'
 import { DatabaseSchema, DatabaseSchemaType } from '../../db/database-schema'
 import RecordProcessor from '../processor'
-import { PrimaryDatabase } from '../../db'
+import { Database } from '../../db'
 import { BackgroundQueue } from '../../background'
 
 const lexId = lex.ids.AppBskyActorProfile
@@ -63,7 +63,7 @@ const notifsForDelete = () => {
 export type PluginType = RecordProcessor<Profile.Record, IndexedProfile>
 
 export const makePlugin = (
-  db: PrimaryDatabase,
+  db: Database,
   background: BackgroundQueue,
 ): PluginType => {
   return new RecordProcessor(db, background, {
