+++ packages/bsky/src/data-plane/server/indexing/plugins/repost.ts
@@ -4,7 +4,7 @@ import { AtUri, normalizeDatetimeAlways } from '@atproto/syntax'
 import * as Repost from '../../../../lexicon/types/app/bsky/feed/repost'
 import * as lex from '../../../../lexicon/lexicons'
 import RecordProcessor from '../processor'
-import { PrimaryDatabase } from '../../db'
+import { Database } from '../../db'
 import { countAll, excluded } from '../../db/util'
 import { DatabaseSchema, DatabaseSchemaType } from '../../db/database-schema'
 import { BackgroundQueue } from '../../background'
@@ -133,7 +133,7 @@ const updateAggregates = async (db: DatabaseSchema, repost: IndexedRepost) => {
 export type PluginType = RecordProcessor<Repost.Record, IndexedRepost>
 
 export const makePlugin = (
-  db: PrimaryDatabase,
+  db: Database,
   background: BackgroundQueue,
 ): PluginType => {
   return new RecordProcessor(db, background, {
