+++ packages/bsky/tests/data-plane/did-cache.test.ts
@@ -81,7 +81,7 @@ describe('did cache', () => {
   })
 
   it('accurately reports expired dids & refreshes the cache', async () => {
-    const didCache = new DidSqlCache(network.bsky.db.getPrimary(), 1, 60000)
+    const didCache = new DidSqlCache(network.bsky.db, 1, 60000)
     const shortCacheResolver = new IdResolver({
       plcUrl: network.bsky.ctx.cfg.didPlcUrl,
       didCache,
@@ -110,7 +110,7 @@ describe('did cache', () => {
   })
 
   it('does not return expired dids & refreshes the cache', async () => {
-    const didCache = new DidSqlCache(network.bsky.db.getPrimary(), 0, 1)
+    const didCache = new DidSqlCache(network.bsky.db, 0, 1)
     const shortExpireResolver = new IdResolver({
       plcUrl: network.bsky.ctx.cfg.didPlcUrl,
       didCache,
